feat: update subscriber overviews
